Clean IPCC categorizations from single-dot comments which are YAML 1.1 numbers; also use title case consistently for the title.
